+++ packages/core/bin/promise-watch.js
@@ -1,0 +1,14 @@
+console.log("PROMISE WATCH");
+
+// const { executeJobs } = require("../dist/execute");
+
+const options = {
+  dir: __dirname,
+};
+
+console.log(options)
+
+// executeJobs(options).catch(err => {
+//   console.error(err);
+//   process.exit(1);
+// });
