+++ packages/core/package.json
@@ -1,10 +1,10 @@
 {
   "name": "@promise-watch/core",
-  "version": "1.0.0",
+  "version": "1.1.0-next",
   "author": "Jason Raimondi <jason@raimondi.us> (https://jasonraimondi.com)",
   "main": "src/index.ts",
   "bin": {
-    "promise-watch": "src/bin/promise-watch.ts"
+    "promise-watch": "bin/promise-watch.js"
   },
   "scripts": {
     "test": "jest",
@@ -29,9 +29,6 @@
   "publishConfig": {
     "main": "dist/index.js",
     "private": false,
-    "license": "MIT",
-    "bin": {
-      "promise-watch": "dist/bin/promise-watch.js"
-    }
+    "license": "MIT"
   }
 }
