--- packages/core/src/bin/promise-watch.ts
@@ -1,1 +1,0 @@
-console.log("PROMISE WATCH")
